Codechange: small changes to domestic advisor.
Help string is now set in the same line as the button is declared (one less place to modify to add new buttons)
Page name is no longer needed when adding buttons as it didn't matter. It just needs to be unique and it's now autogenerated based on index.
Fixed an off by one bug where the far right column could fail to be filled under certain circumstances.