Plotgroups: cleanup, improved speed and more python
Removed all calls to plotgroup update, which were triggered by events, which will not change plotgroups (like plottype changes).
Reduced checks when deleting a plotgroup, which has lost connection with a city.

Added python isConnectedTo() and added list of cities in import/export window
Own cities to the left and foreign cities to the right